+++ gtfs_kit/cleaners.py
@@ -1,7 +1,8 @@
 """
 Functions about cleaning feeds.
 """
-from typing import TYPE_CHECKING, Dict
+from __future__ import annotations
+from typing import TYPE_CHECKING
 
 import pandas as pd
 import numpy as np
@@ -169,7 +170,7 @@ def clean_route_short_names(feed: "Feed") -> "Feed":
 
 def build_aggregate_routes_dict(
     routes: pd.DataFrame, by: str = "route_short_name", route_id_prefix: str = "route_"
-) -> Dict[str, str]:
+) -> dict[str, str]:
     """
     Given a DataFrame of routes, group the routes by route short name, say,
     and assign new route IDs using the given prefix.
@@ -215,28 +216,28 @@ def aggregate_routes(
     routes = feed.routes
     nid_by_oid = build_aggregate_routes_dict(routes, by, route_id_prefix)
 
-    # Update to new route IDs
-    routes["route_id"] = routes["route_id"].map(lambda x: nid_by_oid[x])
+    # Update route IDs in routes
+    routes["route_id"] = routes.route_id.map(lambda x: nid_by_oid[x])
     routes = routes.groupby(by).first().reset_index()
     feed.routes = routes
 
-    # Update route IDs of trips
+    # Update route IDs in trips
     trips = feed.trips
-    trips["route_id"] = trips["route_id"].map(lambda x: nid_by_oid[x])
+    trips["route_id"] = trips.route_id.map(lambda x: nid_by_oid[x])
     feed.trips = trips
 
-    # Update route IDs of transfers
-    if feed.transfers is not None:
-        transfers = feed.transfers
-        transfers["route_id"] = transfers["route_id"].map(lambda x: nid_by_oid[x])
-        feed.transfers = transfers
+    # Update route IDs of fare rules
+    if feed.fare_rules is not None and "route_id" in feed.fare_rules.columns:
+        fr = feed.fare_rules
+        fr["route_id"] = fr.route_id.map(lambda x: nid_by_oid[x])
+        feed.fare_rules = fr
 
     return feed
 
 
 def build_aggregate_stops_dict(
     stops: pd.DataFrame, by: str = "stop_code", stop_id_prefix: str = "stop_"
-) -> Dict[str, str]:
+) -> dict[str, str]:
     """
     Given a DataFrame of stops, group the stops by stop code, say,
     and assign new stop IDs using the given prefix.