+++ gtfs_kit/routes.py
@@ -1,8 +1,8 @@
 """
 Functions about routes.
 """
-from collections import OrderedDict
-from typing import Optional, Iterable, List, Dict, TYPE_CHECKING
+from __future__ import annotations
+from typing import Optional, Iterable, TYPE_CHECKING
 import json
 
 import geopandas as gp
@@ -70,11 +70,11 @@ def compute_route_stats_0(
     - ``'service_duration'``: total of the duration of each trip on
       the route in the given subset of trips; measured in hours
     - ``'service_distance'``: total of the distance traveled by each
-      trip on the route in the given subset of trips; measured in
-      whatever distance units are present in ``trip_stats_subset``;
+      trip on the route in the given subset of trips;
+      measured in kilometers if ``feed.dist_units`` is metric;
+      otherwise measured in miles;
       contains all ``np.nan`` entries if ``feed.shapes is None``
-    - ``'service_speed'``: service_distance/service_duration;
-      measured in distance units per hour
+    - ``'service_speed'``: service_distance/service_duration
     - ``'mean_trip_distance'``: service_distance/num_trips
     - ``'mean_trip_duration'``: service_duration/num_trips
 
@@ -106,7 +106,7 @@ def compute_route_stats_0(
     def compute_route_stats_split_directions(group):
         # Take this group of all trips stats for a single route
         # and compute route-level stats.
-        d = OrderedDict()
+        d = dict()
         d["route_short_name"] = group["route_short_name"].iat[0]
         d["route_type"] = group["route_type"].iat[0]
         d["num_trips"] = group.shape[0]
@@ -147,7 +147,7 @@ def compute_route_stats_0(
         return pd.Series(d)
 
     def compute_route_stats(group):
-        d = OrderedDict()
+        d = dict()
         d["route_short_name"] = group["route_short_name"].iat[0]
         d["route_type"] = group["route_type"].iat[0]
         d["num_trips"] = group.shape[0]
@@ -255,12 +255,13 @@ def compute_route_time_series_0(
       the time bin
     - ``num_trip_ends``: number of trips that end within the
       time bin, ignoring trips that end past midnight
-    - ``service_distance``: sum of the service duration accrued
+    - ``service_distance``: sum of the service distance accrued
+      during the time bin across all trips on the route;
+      measured in kilometers if ``feed.dist_units`` is metric;
+      otherwise measured in miles;
+    - ``service_duration``: sum of the service duration accrued
       during the time bin across all trips on the route;
       measured in hours
-    - ``service_distance``: sum of the service distance accrued
-      during the time bin across all trips on the route; measured
-      in kilometers
     - ``service_speed``: ``service_distance/service_duration``
       for the route
 
@@ -429,7 +430,7 @@ def get_routes(
 def compute_route_stats(
     feed: "Feed",
     trip_stats_subset: pd.DataFrame,
-    dates: List[str],
+    dates: list[str],
     headway_start_time: str = "07:00:00",
     headway_end_time: str = "19:00:00",
     *,
@@ -546,7 +547,7 @@ def build_zero_route_time_series(
 def compute_route_time_series(
     feed: "Feed",
     trip_stats_subset: pd.DataFrame,
-    dates: List[str],
+    dates: list[str],
     freq: str = "5Min",
     *,
     split_directions: bool = False,
@@ -636,7 +637,7 @@ def compute_route_time_series(
 
 
 def build_route_timetable(
-    feed: "Feed", route_id: str, dates: List[str]
+    feed: "Feed", route_id: str, dates: list[str]
 ) -> pd.DataFrame:
     """
     Return a timetable for the given route and dates (YYYYMMDD date strings).
@@ -747,7 +748,7 @@ def routes_to_geojson(
     *,
     split_directions: bool = False,
     include_stops: bool = False,
-) -> Dict:
+) -> dict:
     """
     Return a GeoJSON FeatureCollection of MultiLineString features representing this Feed's routes.
     The coordinates reference system is the default one for GeoJSON,
@@ -760,17 +761,16 @@ def routes_to_geojson(
     If the Feed has no shapes, then raise a ValueError.
     If any of the given route IDs are not found in the feed, then raise a ValueError.
     """
-    if route_ids is not None:
-        D = set(route_ids) - set(feed.routes.route_id)
-        if D:
-            raise ValueError(f"Route IDs {D} not found in feed.")
+    if route_ids is None or not list(route_ids):
+        route_ids = feed.routes.route_id
+
+    D = set(route_ids) - set(feed.routes.route_id)
+    if D:
+        raise ValueError(f"Route IDs {D} not found in feed.")
 
     # Get routes
     g = geometrize_routes(feed, route_ids=route_ids, split_directions=split_directions)
-    if g.empty:
-        collection = {"type": "FeatureCollection", "features": []}
-    else:
-        collection = json.loads(g.to_json())
+    collection = json.loads(g.to_json())
 
     # Get stops if desired
     if include_stops:
@@ -792,7 +792,7 @@ def routes_to_geojson(
 def map_routes(
     feed: "Feed",
     route_ids: Iterable[str],
-    color_palette: List[str] = cs.COLORS_SET2,
+    color_palette: list[str] = cs.COLORS_SET2,
     *,
     include_stops: bool = False,
 ):